:bug: Fix: resolve sign in auth data management logic error

+++ authentication/authentication-data/src/main/java/com/sowhat/authentication_data/model/response/SignInResult.kt
@@ -11,5 +11,5 @@ data class SignInResult(
     @SerialName("isJoined")
     val isJoined: Boolean,
     @SerialName("memberId")
-    val memberId: Int?
+    val memberId: Long?
 )
